requirements.txt
black==22.10.0
psutil==5.9.4
python-dotenv==0.21.0
rich==12.6.0